+++ lib/image/ImageHeader.cpp
@@ -270,7 +270,8 @@ void ImageHeader::adjustComponents(const ComponentType& componentType, uint32_t 
   m_fileComponentType = componentType;
   m_fileComponentTypeAsString = compString;
   m_fileComponentSizeInBytes = sizeInBytes;
-  m_fileImageSizeInBytes = m_fileComponentSizeInBytes * m_numComponentsPerPixel * m_numPixels;
+  m_fileImageSizeInBytes =
+    static_cast<uint64_t>(m_fileComponentSizeInBytes) * m_numComponentsPerPixel * m_numPixels;
 
   m_memoryComponentType = m_fileComponentType;
   m_memoryComponentTypeAsString = m_fileComponentTypeAsString;
@@ -313,8 +314,10 @@ void ImageHeader::setNumComponentsPerPixel(uint32_t numComponents)
   }
 
   m_numComponentsPerPixel = numComponents;
-  m_fileImageSizeInBytes = m_fileComponentSizeInBytes * m_numComponentsPerPixel * m_numPixels;
-  m_memoryImageSizeInBytes = m_memoryComponentSizeInBytes * m_numComponentsPerPixel * m_numPixels;
+  m_fileImageSizeInBytes =
+    static_cast<uint64_t>(m_fileComponentSizeInBytes) * m_numComponentsPerPixel * m_numPixels;
+  m_memoryImageSizeInBytes =
+    static_cast<uint64_t>(m_memoryComponentSizeInBytes) * m_numComponentsPerPixel * m_numPixels;
 
   /// @note \c m_ioInfoInMemory has not been updated. It is never retrieved by the client or used
   /// later. m_ioInfoInMemory.m_pixelInfo.m_numComponents = m_numComponentsPerPixel;
